test: improve coverage and relocate command tests
